+++ packages/wasm-crypto-init/src/wasm.ts
@@ -14,4 +14,4 @@ export { packageInfo } from './packageInfo';
  * @description
  * Creates an interface using only WASM
  */
-export const createWasm: InitFn<WasmCryptoInstance> = createWasmFn('crypto', wasmBytes, null);
+export const createWasm: InitFn<WasmCryptoInstance> = /*#__PURE__*/ createWasmFn('crypto', wasmBytes, null);