+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # CHANGELOG
 
+## master
+
+Changes:
+
+- Add `/*#__PURE__*/` annotations for specific `export const something = someFunction(...)
+
+
 ## 6.3.1 Jul 21, 2022
 
 Changes: